Move open_checkpoint from static_ct_api to tlog_tiles; remove static_ct_api dep from old ct_worker
