+++ crates/tlog_tiles/Cargo.toml
@@ -25,6 +25,7 @@ crate-type = ["rlib"]
 base64.workspace = true
 ed25519-dalek.workspace = true
 length_prefixed.workspace = true
+rand.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 sha2.workspace = true
